Move browse interface to root URL
